requirements.txt
# Core dependencies
streamlit==1.29.0
python-dotenv==1.0.0
pymongo==4.6.1

# Image processing and medical imaging
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pydicom==2.4.4
nibabel==5.2.0

# AI and ML
openai==1.3.0
requests==2.31.0

# Scientific and data processing
biopython==1.83
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Report generation
reportlab==4.0.7
markdown==3.5.1

# Web framework (optional - for FastAPI if needed)
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
